requirements.txt
pip>=9.0.1
Flask==0.12.2
cassandra-driver